File structure and new compoent wrappers
