+++ src/components/CompletionView/src/Tabs/TreeTab.tsx
@@ -5,6 +5,5 @@ import { useDataContext } from '../Context/DataProvider';
 export const TreeTab = (): JSX.Element => {
     const { data } = useFilteredData();
     const { treeOptions } = useDataContext();
-
     return <Tree data={data} treeOptions={treeOptions} />;
 };
