+++ src/components/Routes/ComponentWrapper.tsx
@@ -1,0 +1,11 @@
+import { AppManifest } from '../../apps/apps';
+import useClientContext from '../../context/clientContext';
+import { DefaultRouteComponent } from './DefaultRouteComponent';
+
+export function ComponentWrapper(route: AppManifest): JSX.Element {
+    const { appConfig, authProvider } = useClientContext();
+    const Component = route.app?.component || DefaultRouteComponent;
+    const api = { ...route, authProvider, appConfig };
+
+    return <Component {...api} />;
+}
