+++ src/components/Routes/Routes.tsx
@@ -1,65 +1,15 @@
-import { Icon } from '@equinor/eds-core-react';
-import { tokens } from '@equinor/eds-tokens';
-import { Link, Route, Routes } from 'react-router-dom';
-import { AppGroupe, AppManifest, Apps } from '../../apps/apps';
+import { Route, Routes } from 'react-router-dom';
+import { Apps } from '../../apps/apps';
 import { useApps } from '../../apps/useApps';
 import useClientContext from '../../context/clientContext';
+import { DataView } from '../CompletionView/src/DataView';
 import { HomePage } from '../HomePage/HomePage';
-import { DefaultRouteComponent } from './DefaultRouteComponent';
-
-function ComponentWrapper(route: AppManifest) {
-    const { appConfig, authProvider } = useClientContext();
-    const Component = route.app?.component || DefaultRouteComponent;
-    route.app?.setup &&
-        route.app.setup({
-            ...route,
-            appConfig,
-            authProvider,
-        });
-
-    const api = { ...route, authProvider, appConfig };
-
-    return <Component {...api} />;
-}
-
-interface GroupComponentWrapperProps extends AppGroupe {
-    links: AppManifest[];
-    groupeId: string;
-}
-
-function GroupComponentWrapper(group: GroupComponentWrapperProps) {
-    return (
-        <div>
-            This is a AppGroupe
-            <h2>{group.name}</h2>
-            <div style={{ display: 'flex', flexDirection: 'column' }}>
-                {group.links.map((item) => {
-                    const CustomIcon = item.icon;
-                    return (
-                        <div key={`group-link-${item.shortName}`}>
-                            <Link className="link" to={`/${group.groupeId}/${item.shortName}`}>
-                                {CustomIcon && typeof CustomIcon !== 'string' && <CustomIcon />}
-
-                                {CustomIcon && typeof CustomIcon === 'string' && (
-                                    <Icon
-                                        name={CustomIcon}
-                                        title={item.title}
-                                        color={tokens.colors.text.static_icons__secondary.rgba}
-                                    />
-                                )}
-
-                                {<span>{item.title}</span>}
-                            </Link>
-                        </div>
-                    );
-                })}
-            </div>
-        </div>
-    );
-}
+import { ComponentWrapper } from './ComponentWrapper';
+import { GroupComponentWrapper } from './GroupComponentWrapper';
 
 export function ClientRoutes(): JSX.Element {
     const { apps, appGroups } = useApps();
+    const { appConfig, authProvider } = useClientContext();
 
     return (
         <Routes>
@@ -81,18 +31,27 @@ export function ClientRoutes(): JSX.Element {
                 );
             })}
             {apps.map((route) => {
+                route.app?.setup &&
+                    route.app.setup({
+                        ...route,
+                        appConfig,
+                        authProvider,
+                    });
+
+                //  TODO extend this with new apptype
                 if (route.app?.appType === 'DataViewer') {
+                    const api = { ...route, authProvider, appConfig };
                     return (
                         <>
                             <Route
                                 key={route.shortName}
                                 path={`${route.groupe.toString()}/${route.shortName}`}
-                                element={<ComponentWrapper {...route} />}
+                                element={<DataView {...api} />}
                             />
                             <Route
                                 key={route.shortName + 'id'}
                                 path={`${route.groupe.toString()}/${route.shortName}/:id`}
-                                element={<ComponentWrapper {...route} />}
+                                element={<DataView {...api} />}
                             />
                         </>
                     );
@@ -105,7 +64,6 @@ export function ClientRoutes(): JSX.Element {
                     />
                 );
             })}
-            {/* <Route element={<Redirect to="/" />} /> */}
         </Routes>
     );
 }
